+++ pyro/poutine/escape_messenger.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from .util import NonlocalExit
 
-from .poutine import Messenger
+from .messenger import Messenger
 
 
 class EscapeMessenger(Messenger):
