+++ pyro/poutine/lift_messenger.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 from pyro import params
 from pyro.distributions import Distribution
 
-from .poutine import Messenger
+from .messenger import Messenger
 
 
 class LiftMessenger(Messenger):
