+++ pyro/poutine/trace_messenger.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from .poutine import Messenger, Handler
+from .messenger import Messenger, Handler
 from .trace import Trace
 from .util import site_is_subsample
